api/pyproject.toml
[project]
name = "django-template-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Nguyễn Hoàng Quân",email = "quannguyen5@beeketing.net"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "graphene-django (>=3.2.3,<4.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
]

[tool.poetry]
packages = [{include = "django_template_project", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<4"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.migrations
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["config", "example_app"]
skip_gitignore = true